+++ CHANGELOG.md
@@ -1,5 +1,6 @@
 ## **Unreleased**
 **Hearthstone**:
+- Fixed an issue where Reach Equilibrium (Priest's Quest) would count as 3 spells in the spell counter.
 - Fixed an issue where Zilliax would always display as 0 mana on deck image exporter.
 
 ## **Release v1.46.1 - 2025-07-23**